refactor: split main.go in individual files per endpoint/handler. Also moved tests to corresponding files.
